Extract mock::consensus_state mod
