+++ modules/src/mock/client_def.rs
@@ -18,7 +18,8 @@ use crate::core::ics23_commitment::merkle::apply_prefix;
 use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
 use crate::core::ics24_host::path::ClientConsensusStatePath;
 use crate::core::ics24_host::Path;
-use crate::mock::client_state::{MockClientState, MockConsensusState};
+use crate::mock::client_state::MockClientState;
+use crate::mock::consensus_state::MockConsensusState;
 use crate::mock::header::MockHeader;
 use crate::prelude::*;
 use crate::Height;
