+++ modules/src/mock/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod client_def;
 pub mod client_state;
+pub mod consensus_state;
 pub mod context;
 pub mod header;
 pub mod host;
