+++ modules/tests/runner/mod.rs
@@ -27,7 +27,8 @@ use ibc::core::ics23_commitment::commitment::{CommitmentPrefix, CommitmentProofB
 use ibc::core::ics24_host::identifier::{ChainId, ClientId, ConnectionId};
 use ibc::core::ics26_routing::error as routing_error;
 use ibc::core::ics26_routing::msgs::Ics26Envelope;
-use ibc::mock::client_state::{MockClientState, MockConsensusState};
+use ibc::mock::client_state::MockClientState;
+use ibc::mock::consensus_state::MockConsensusState;
 use ibc::mock::context::MockContext;
 use ibc::mock::header::MockHeader;
 use ibc::mock::host::HostType;
